[JVM_IR] Improve debugging of default lambda parameters
CallableReferenceLowering was omitting source offsets for the
expression of the default parameter in the case it was a lambda,
meaning that no line numbers where present in the resulting
bytecode. This in turn caused incorrect debugging behaviour when
inlining the $default handler.

TODO: Nested inline functions causes incorrect source maps due to the
nature of InlineCodegen.